+++ visualization/app/codeCharta/state/store/state.actions.ts
@@ -3,6 +3,7 @@ import { defaultAppSettings } from "./appSettings/appSettings.actions"
 import { defaultFileSettings } from "./fileSettings/fileSettings.actions"
 import { defaultDynamicSettings } from "./dynamicSettings/dynamicSettings.actions"
 import { defaultTreeMapSettings } from "./treeMap/treeMap.actions"
+import { defaultFiles } from "./files/files.actions"
 
 export enum StateActions {
 	SET_STATE = "SET_STATE"
@@ -26,5 +27,6 @@ export const defaultState: State = {
 	appSettings: defaultAppSettings,
 	fileSettings: defaultFileSettings,
 	dynamicSettings: defaultDynamicSettings,
-	treeMap: defaultTreeMapSettings
+	treeMap: defaultTreeMapSettings,
+	files: defaultFiles
 }