+++ visualization/app/codeCharta/ui/legendPanel/legendPanel.component.spec.ts
@@ -2,10 +2,8 @@ import "./legendPanel.module"
 
 import { LegendPanelController, PackageList } from "./legendPanel.component"
 import { instantiateModule, getService } from "../../../../mocks/ng.mockhelper"
-import { FileStateService } from "../../state/fileState.service"
 import { IRootScopeService } from "angular"
-import { CCFile, ColorRange } from "../../codeCharta.model"
-import { TEST_FILE_DATA } from "../../util/dataMocks"
+import { ColorRange } from "../../codeCharta.model"
 import _ from "lodash"
 import { StoreService } from "../../state/store.service"
 import { ColorRangeService } from "../../state/store/dynamicSettings/colorRange/colorRange.service"
@@ -19,9 +17,6 @@ describe("LegendPanelController", () => {
 	let legendPanelController: LegendPanelController
 	let $rootScope: IRootScopeService
 	let storeService: StoreService
-	let fileStateService: FileStateService
-
-	let file: CCFile
 
 	beforeEach(() => {
 		restartSystem()
@@ -33,13 +28,10 @@ describe("LegendPanelController", () => {
 
 		$rootScope = getService<IRootScopeService>("$rootScope")
 		storeService = getService<StoreService>("storeService")
-		fileStateService = getService<FileStateService>("fileStateService")
-
-		file = _.cloneDeep(TEST_FILE_DATA)
 	}
 
 	function rebuildController() {
-		legendPanelController = new LegendPanelController($rootScope, storeService, fileStateService)
+		legendPanelController = new LegendPanelController($rootScope, storeService)
 	}
 
 	describe("constructor", () => {
@@ -93,16 +85,12 @@ describe("LegendPanelController", () => {
 	})
 
 	describe("onMarkedPackagesChanged", () => {
-		beforeEach(() => {
-			fileStateService["fileStates"].push({ file, selectedAs: null })
-		})
-
 		it("set correct markingPackage in Legend", () => {
-			const markedPackages = [{ color: "#FF0000", path: "/root", attributes: {} }]
+			const markedPackages = [{ color: "#FF0000", path: "/root" }]
 			const expectedPackageLists: PackageList[] = [
 				{
 					colorPixel: "data:image/gif;base64,R0lGODlhAQABAPAAAP8AAP///yH5BAAAAAAALAAAAAABAAEAAAICRAEAOw==",
-					markedPackages: [{ color: "#FF0000", path: "/root", attributes: { name: "/root" } }]
+					markedPackages: [{ color: "#FF0000", path: "/root" }]
 				}
 			]
 
@@ -111,21 +99,6 @@ describe("LegendPanelController", () => {
 			expect(legendPanelController["_viewModel"].packageLists).toEqual(expectedPackageLists)
 		})
 
-		it("shorten too long pathName in middle of the string for legendPanel", () => {
-			const markedPackages = [{ color: "#FF0000", path: "/root/a/longNameToBeShortenedInLegend", attributes: {} }]
-			const shortenedPathname = "longNameToBe...enedInLegend"
-
-			legendPanelController.onMarkedPackagesChanged(markedPackages)
-			expect(legendPanelController["_viewModel"].packageLists[0].markedPackages[0].attributes["name"]).toEqual(shortenedPathname)
-		})
-
-		it("shorten too long pathName at beginning of the string for legendPanel", () => {
-			const markedPackages = [{ color: "#FF0000", path: "/root/a/andAnotherLongNameToShorten", attributes: {} }]
-			const shortenedPathname = ".../andAnotherLongNameToShorten"
-
-			legendPanelController.onMarkedPackagesChanged(markedPackages)
-			expect(legendPanelController["_viewModel"].packageLists[0].markedPackages[0].attributes["name"]).toEqual(shortenedPathname)
-		})
 		it("should update the ColorRange when it is changed", () => {
 			const newColorRange: ColorRange = { from: 13, to: 33 }