+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 ### Fixed 🐞
 
 - Loading Gif not displayed when preparing to render a new map #857
+- Selecting zero files in Multiple mode will not trigger the 3D CodeMap creation
 
 ### Chore 👨‍💻 👩‍💻
 
@@ -31,6 +32,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 - Replaced non standard `[[` in sh scripts #849
 - Improved performance for loading a new file #836
+- Marked Packages are loaded from files #798
 
 ### Chore 👨‍💻 👩‍💻